+++ front/pages/api/stripe/webhook.ts
@@ -1,4 +1,3 @@
-import sgMail from "@sendgrid/mail";
 import { NextApiRequest, NextApiResponse } from "next";
 import { pipeline, Writable } from "stream";
 import Stripe from "stripe";
@@ -17,12 +16,14 @@ import { PlanInvitation } from "@app/lib/models/plan";
 import { generateModelSId } from "@app/lib/utils";
 import logger from "@app/logger/logger";
 import { apiError, withLogging } from "@app/logger/withlogging";
+import {
+  cancelMessage,
+  getAdminEmailsForWorkspace,
+  reactivateMessage,
+  sendEmail,
+} from "@app/lib/email";
 
-const {
-  SENDGRID_API_KEY = "",
-  STRIPE_SECRET_KEY = "",
-  STRIPE_SECRET_WEBHOOK_KEY = "",
-} = process.env;
+const { STRIPE_SECRET_KEY = "", STRIPE_SECRET_WEBHOOK_KEY = "" } = process.env;
 
 export type GetResponseBody = {
   success: true;
@@ -334,9 +335,9 @@ async function handler(
             }
             // send email to admins
             for (const adminEmail of adminEmails)
-              await sendAdminEmail(
+              await sendEmail(
                 adminEmail,
-                endDate ? cancellationMessage(endDate) : reactivationMessage
+                endDate ? cancelMessage(endDate) : reactivateMessage
               );
           }
           break;
@@ -386,89 +387,4 @@ async function handler(
   }
 }
 
-async function getAdminEmailsForWorkspace(id: ModelId) {
-  const users = await User.findAll({
-    attributes: ["email"],
-    include: [
-      {
-        model: Membership,
-        required: true,
-        where: {
-          role: "admin",
-        },
-        attributes: [],
-        include: [
-          {
-            model: Workspace,
-            required: true,
-            attributes: [],
-            where: { id },
-          },
-        ],
-      },
-    ],
-  });
-  return users.map((u) => u.email);
-}
-
-sgMail.setApiKey(SENDGRID_API_KEY);
-
-const cancellationMessage = (date: Date) => {
-  const options: Intl.DateTimeFormatOptions = {
-    year: "numeric",
-    month: "short",
-    day: "numeric",
-  };
-  const formattedDate = date.toLocaleDateString("en-US", options);
-  return {
-    from: {
-      name: "Dust team",
-      email: "team@dust.tt",
-    },
-    subject: `[Dust] Your subscription has been canceled; important information`,
-    text: `<p>You have requested to cancel your subscription</p> 
-  <p>Your subscription will end at the end of your current billing period (${formattedDate}). You can reactivate your subscription at any time before then. If you do not reactivate your subscription, you will be switched back to our free plan at the end of the current billing period.</p>
-  <p>This will have the following consequences:</p>
-  <ul>
-  <li>all users except one will be removed from your workspace;</li>
-  <li>your connections will be deleted</li>
-  <li>the rest of your data (conversations, custom assistants, custom datasource) will still be accessible but with multiple limitations.</li>
-  <li>you will be subjected to the <a href="https://dust.tt/#sectionPricing">restrictions of the free plan</a></li>
-  </ul>
-  <p>Complete details as to what will happen are available at <a href="https://dust-tt.notion.site/What-happens-when-we-cancel-our-Dust-subscription-59aad3866dcc4bbdb26a54e1ce0d848a?pvs=4">subscription cancelling FAQ</a>.</p>
-  <p>If you have any questions, please contact us at team@dust.tt.</p>
-  <p>Best,</p>
-  <p>The Dust team</p>`,
-  };
-};
-
-const reactivationMessage = {
-  from: {
-    name: "Dust team",
-    email: "team@dust.tt",
-  },
-  subject: `[Dust] Your subscription has been reactivated`,
-  text: `<p>You have requested to reactivate your subscription.</p> 
-  <p>Therefore, your subscription will not be canceled at the end of the billing period, no downgrade actions will take place, and you can continue using Dust as usual.</p>
-  <p>We really appreciate you renewing your trust in us.</p>
-  <p>If you have any questions, we'll gladly answer at team@dust.tt.</p>
-  <p>Best,</p>
-  <p>The Dust team</p>`,
-};
-
-async function sendAdminEmail(email: string, message: any) {
-  const msg = { ...message, to: email };
-  try {
-    await sgMail.send(msg);
-    logger.info(
-      { email, subject: message.subject },
-      "Sending email to admin about subscription."
-    );
-  } catch (error) {
-    logger.error(
-      { error, email },
-      "Error sending email to admin about subscription cancellation."
-    );
-  }
-}
 export default withLogging(handler);
